Moving TeleopDrive
